restaurantlist component functional within app, bug incorporating it to router in main.tsx
