+++ src/App.tsx
@@ -2,6 +2,7 @@ import Icon from './components/Icon'
 import Button from './components/Button'
 import './App.css'
 import Bounded from './components/Bounded'
+import RestaurantList from './components/RestaurantList'
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
         to={"X"}>
           OK
         </Button>
+        <RestaurantList />
       </Bounded>
     </>
   )
